tenant store and tested

+++ management/domain/tenant_repo.go
@@ -9,7 +9,7 @@ type TenantRepo interface {
 	//Find by id or name
 	FindByIdOrName(ctx context.Context,idOrName string)(*Tenant,error)
 	//total count
-	GetCount(ctx context.Context,)(int,error)
+	GetCount(ctx context.Context,)(int64,error)
 	//get paged list
 	GetPaged(ctx context.Context,p common.Pagination)(c int64,t []*Tenant,err error)
 	//create a tenant
